UI: Add archived and sent tabs to the inbox
This change adds a tab to the inbox that shows archived notifications
and a tab that shows your sent messages.

Change-type: patch
